Add support for line-based messages
